+++ includes/frontend/class-post.php
@@ -6,7 +6,7 @@
  * @copyright  WebMan Design, Oliver Juhas
  *
  * @since    2.0.0
- * @version  2.0.0
+ * @version  2.3.0
  *
  * Contents:
  *
@@ -159,7 +159,7 @@ class Modern_Post {
 		 * @uses  `wmhook_modern_title_primary_disable` global hook to disable `#primary` section H1
 		 *
 		 * @since    2.0.0
-		 * @version  2.0.0
+		 * @version  2.3.0
 		 *
 		 * @param  array $args Heading setup arguments
 		 */
@@ -173,7 +173,7 @@ class Modern_Post {
 
 				if ( false !== $pre ) {
 					if ( true !== $pre ) {
-						echo $pre;
+						echo $pre; // Functionality bypass via filter.
 					}
 					return;
 				}
@@ -340,7 +340,7 @@ class Modern_Post {
 		 * Post navigation
 		 *
 		 * @since    2.0.0
-		 * @version  2.0.0
+		 * @version  2.3.0
 		 */
 		public static function navigation() {
 
@@ -427,7 +427,7 @@ class Modern_Post {
 				}
 
 				if ( $styles ) {
-					$styles = (string) apply_filters( 'wmhook_modern_post_navigation_styles', '<style id="post-navigation-css" type="text/css">' . apply_filters( 'wmhook_modern_esc_css', $styles ) . '</style>' );
+					$styles = (string) apply_filters( 'wmhook_modern_post_navigation_styles', '<style id="post-navigation-css" type="text/css">' . Modern_Customize_Styles::esc_css( $styles ) . '</style>' );
 				}
 
 
@@ -444,13 +444,10 @@ class Modern_Post {
 
 
 			/**
-			 * Post navigation styles
-			 *
-			 * @uses  `wmhook_modern_inline_styles_handle` global hook
-			 * @uses  `wmhook_modern_esc_css` global hook
+			 * Post navigation styles.
 			 *
 			 * @since    1.0.0
-			 * @version  2.0.0
+			 * @version  2.3.0
 			 */
 			public static function navigation_styles() {
 
@@ -500,8 +497,8 @@ class Modern_Post {
 					if ( ! empty( $output ) ) {
 
 						wp_add_inline_style(
-							(string) apply_filters( 'wmhook_modern_inline_styles_handle', 'modern-stylesheet-global' ),
-							(string) apply_filters( 'wmhook_modern_esc_css', $output, 'Modern_Post::navigation_styles' )
+							'modern',
+							Modern_Customize_Styles::esc_css( $output, 'Modern_Post::navigation_styles' )
 						);
 
 					}
@@ -551,18 +548,18 @@ class Modern_Post {
 				 * fine with customizer options.
 				 *
 				 * @since    2.0.0
-				 * @version  2.0.0
+				 * @version  2.2.0
 				 */
 				public static function template_front_display_blog() {
 
 					// Helper variables
 
-						$location = explode( '|', (string) get_theme_mod( 'layout_location_front_blog', 'tha_content_before|20' ) );
+						$location = explode( '|', (string) Modern_Library_Customize::get_theme_mod( 'layout_location_front_blog' ) );
 						if ( ! isset( $location[1] ) ) {
 							$location[1] = 10;
 						}
 
-						if ( 1 > intval( get_theme_mod( 'layout_posts_per_page_front_blog', 6 ) ) ) {
+						if ( 1 > intval( Modern_Library_Customize::get_theme_mod( 'layout_posts_per_page_front_blog' ) ) ) {
 							$location[0] = false;
 						}