+++ library/includes/classes/class-customize-control-radio-matrix.php
@@ -4,11 +4,14 @@
  *
  * Customizer matrix radio fields.
  *
- * @package     WebMan WordPress Theme Framework
  * @subpackage  Customize
  *
+ * @package    WebMan WordPress Theme Framework
+ * @copyright  WebMan Design, Oliver Juhas
+ *
  * @since    1.0.0
- * @version  2.1.0
+ * @version  2.7.0
+ * @version  2.3.0
  */
 class Modern_Customize_Control_Radio_Matrix extends WP_Customize_Control {
 
@@ -25,12 +28,12 @@ class Modern_Customize_Control_Radio_Matrix extends WP_Customize_Control {
 			// Scripts
 
 				wp_enqueue_script(
-						'modern-customize-control-radio-matrix',
-						get_theme_file_uri( MODERN_LIBRARY_DIR . 'js/customize-control-radio-matrix.js' ),
-						array( 'customize-controls' ),
-						esc_attr( MODERN_THEME_VERSION ),
-						true
-					);
+					'modern-customize-control-radio-matrix',
+					get_theme_file_uri( MODERN_LIBRARY_DIR . 'js/customize-control-radio-matrix.js' ),
+					array( 'customize-controls' ),
+					MODERN_THEME_VERSION,
+					true
+				);
 
 	} // /enqueue
 
@@ -44,10 +47,10 @@ class Modern_Customize_Control_Radio_Matrix extends WP_Customize_Control {
 
 				?>
 
-				<span class="customize-control-title"><?php echo $this->label; ?></span>
-				<?php if ( $this->description ) : ?><span class="description customize-control-description"><?php echo $this->description; ?></span><?php endif; ?>
+				<span class="customize-control-title"><?php echo esc_html( $this->label ); ?></span>
+				<?php if ( $this->description ) : ?><span class="description customize-control-description"><?php echo wp_kses_post( $this->description ); ?></span><?php endif; ?>
 
-				<div class="<?php echo trim( 'custom-radio-container ' . $this->class ); ?>">
+				<div class="<?php echo esc_attr( trim( 'custom-radio-container ' . $this->class ) ); ?>">
 					<?php
 
 					$i = 0;
@@ -66,14 +69,14 @@ class Modern_Customize_Control_Radio_Matrix extends WP_Customize_Control {
 
 						?>
 
-						<label for="<?php echo esc_attr( $this->id . ++$i ); ?>"<?php echo $active_class . $title; ?>>
-							<?php echo $name; ?>
-							<input class="custom-radio-item" type="radio" value="<?php echo esc_attr( $value ); ?>" name="<?php echo esc_attr( $this->id ); ?>" id="<?php echo esc_attr( $this->id . $i ); ?>" <?php echo $this->get_link() . $checked; ?> />
+						<label for="<?php echo esc_attr( $this->id . ++$i ); ?>"<?php echo $active_class . $title; /* WPCS: XSS OK. */ ?>>
+							<?php echo wp_kses_post( $name ); ?>
+							<input class="custom-radio-item" type="radio" value="<?php echo esc_attr( $value ); ?>" name="<?php echo esc_attr( $this->id ); ?>" id="<?php echo esc_attr( $this->id . $i ); ?>" <?php echo $this->get_link() . $checked; /* WPCS: XSS OK. */ ?> />
 						</label>
 
 						<?php
 
-					} // /foreach
+					}
 
 					?>
 				</div>