+++ library/js/customize-control-radio-matrix.js
@@ -3,13 +3,19 @@
  *
  * Customizer matrix radio fields.
  *
- * @package     WebMan WordPress Theme Framework
  * @subpackage  Customize
  *
+ * @package    WebMan WordPress Theme Framework
+ * @copyright  WebMan Design, Oliver Juhas
+ *
  * @since    1.0.0
- * @version  2.2.0
+ * @version  2.7.0
+ * @version  2.3.0
  */
 ( function( exports, $ ) {
+
+	'use strict';
+
 	$( wp.customize ).on( 'ready', function() {